+++ src/utils/logger.ts
@@ -4,7 +4,7 @@ import pino = require('pino');
  * BASE LOGS *
  *************/
 
-const baseLogger = pino();
+const baseLogger = pino({customLevels: { verbose: 25, silly: 5 }});
 
 function getBaseLogger() {
   return baseLogger;